Extract coloring logic to Color helper module
